restructure interpolation and add extrapolation
