+++ fastsim-core/src/utils/interp/two.rs
@@ -1,0 +1,132 @@
+//! 2-dimensional interpolation
+
+use super::*;
+
+#[derive(Debug)]
+pub struct Interp2D {
+    pub x: Vec<f64>,
+    pub y: Vec<f64>,
+    pub f_xy: Vec<Vec<f64>>,
+    pub strategy: Strategy,
+    pub extrapolate: Extrapolate,
+    _phantom: PhantomData<()>, // phantom private field to prevent direct instantiation in other modules
+}
+
+impl Interp2D {
+    pub fn new(
+        x: Vec<f64>,
+        y: Vec<f64>,
+        f_xy: Vec<Vec<f64>>,
+        strategy: Strategy,
+        extrapolate: Extrapolate,
+    ) -> anyhow::Result<Self> {
+        let interp = Self {
+            x,
+            y,
+            f_xy,
+            strategy,
+            extrapolate,
+            _phantom: PhantomData,
+        };
+        interp.validate()?;
+        Ok(interp)
+    }
+
+    pub fn linear(&self, point: &[f64]) -> anyhow::Result<f64> {
+        let x_l = find_nearest_index(&self.x, point[0])?;
+        let x_u = x_l + 1;
+        let x_diff = (point[0] - self.x[x_l]) / (self.x[x_u] - self.x[x_l]);
+
+        let y_l = find_nearest_index(&self.y, point[1])?;
+        let y_u = y_l + 1;
+        let y_diff = (point[1] - self.y[y_l]) / (self.y[y_u] - self.y[y_l]);
+
+        // interpolate in the x-direction
+        let c0 = self.f_xy[x_l][y_l] * (1.0 - x_diff) + self.f_xy[x_u][y_l] * x_diff;
+        let c1 = self.f_xy[x_l][y_u] * (1.0 - x_diff) + self.f_xy[x_u][y_u] * x_diff;
+
+        // interpolate in the y-direction
+        Ok(c0 * (1.0 - y_diff) + c1 * y_diff)
+    }
+}
+
+impl InterpMethods for Interp2D {
+    fn validate(&self) -> anyhow::Result<()> {
+        let x_grid_len = self.x.len();
+        let y_grid_len = self.y.len();
+
+        // Check that each grid dimension has elements
+        ensure!(
+            x_grid_len != 0 && y_grid_len != 0,
+            "Supplied grid coordinates cannot be empty"
+        );
+        // Check that grid points are monotonically increasing
+        ensure!(
+            self.x.windows(2).all(|w| w[0] < w[1]) && self.y.windows(2).all(|w| w[0] < w[1]),
+            "Supplied coordinates must be sorted and non-repeating"
+        );
+        // Check that grid and values are compatible shapes
+        let x_dim_ok = x_grid_len == self.f_xy.len();
+        let y_dim_ok = self
+            .f_xy
+            .iter()
+            .map(|y_vals| y_vals.len())
+            .all(|y_val_len| y_val_len == y_grid_len);
+        ensure!(
+            x_dim_ok && y_dim_ok,
+            "Supplied grid and values are not compatible shapes"
+        );
+
+        Ok(())
+    }
+
+    fn interpolate(&self, point: &[f64]) -> anyhow::Result<f64> {
+        match self.strategy {
+            Strategy::Linear => self.linear(point),
+            _ => bail!(
+                "Provided strategy {:?} is not applicable for 2-D interpolation",
+                self.strategy
+            ),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_linear() {
+        let x = vec![0.05, 0.10, 0.15];
+        let y = vec![0.10, 0.20, 0.30];
+        let f_xy = vec![vec![0., 1., 2.], vec![3., 4., 5.], vec![6., 7., 8.]];
+        let interp = Interpolator::Interp2D(
+            Interp2D::new(
+                x.clone(),
+                y.clone(),
+                f_xy.clone(),
+                Strategy::Linear,
+                Extrapolate::Error,
+            )
+            .unwrap(),
+        );
+        assert_eq!(interp.interpolate(&[x[2], y[1]]).unwrap(), 7.);
+        assert_eq!(interp.interpolate(&[x[2], y[1]]).unwrap(), 7.);
+    }
+
+    #[test]
+    fn test_linear_offset() {
+        let interp = Interpolator::Interp2D(
+            Interp2D::new(
+                vec![0., 1.],
+                vec![0., 1.],
+                vec![vec![0., 1.], vec![2., 3.]],
+                Strategy::Linear,
+                Extrapolate::Error,
+            )
+            .unwrap(),
+        );
+        let interp_res = interp.interpolate(&[0.25, 0.65]).unwrap();
+        assert_eq!(interp_res, 1.1500000000000001) // 1.15
+    }
+}
